+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.0.2] - 2024-06-26
 - Mocking Kafka publish in PersonWrite tests
-- 
+- Initial Kafka Setup in PersonWrite
 
 ## [0.0.1] - 2024-06-26